+++ internal/sidekick/surfer/command_writer_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package gcloud
+package surfer
 
 import (
 	"errors"
@@ -24,7 +24,7 @@ import (
 	"time"
 
 	"github.com/google/go-cmp/cmp"
-	"github.com/googleapis/librarian/internal/sidekick/gcloud/declarative"
+	"github.com/googleapis/librarian/internal/sidekick/surfer/declarative"
 )
 
 func TestWritePartialFile_Output(t *testing.T) {
