+++ internal/sidekick/surfer/declarative/declarative.go
