+++ internal/sidekick/surfer/provider/config_test.go
