+++ internal/sidekick/surfer/provider/loader.go
