+++ internal/sidekick/surfer/provider/loader_test.go
