+++ internal/sidekick/surfer/provider/resource.go
