+++ internal/sidekick/surfer/provider/resource_test.go
