+++ internal/sidekick/surfer/provider/service_info.go
